feat(backend): make bun backend as a setting of npm backend
