+++ e2e/backend/test_npm_bun
@@ -1,0 +1,13 @@
+#!/usr/bin/env bash
+
+# 
+cat >mise.toml <<EOF
+[settings.npm]
+bun = true
+EOF
+
+mise use bun
+
+assert "mise x npm:prettier@3.1.0 -- prettier -v" "3.1.0"
+assert "FORCE_COLOR=0 mise x npm:@antfu/ni@0.21.12 -- ni -v 2>/dev/null | head -n1" "@antfu/ni  v0.21.12"
+assert_succeed "mise install npm:tldr@3.4.0"
