+++ src/cli/settings/ls.rs
@@ -102,6 +102,9 @@ mod tests {
 
         [node]
 
+        [npm]
+        bun = true
+
         [pipx]
         uvx = false
 
@@ -158,6 +161,8 @@ mod tests {
         lockfile
         node
         not_found_auto_install
+        npm
+        npm.bun
         paranoid
         pin
         pipx
