+++ src/cli/settings/set.rs
@@ -168,6 +168,9 @@ pub mod tests {
 
         [node]
 
+        [npm]
+        bun = true
+
         [pipx]
         uvx = false
